[chore] Move `//index/internal/collections' to `//internal/collections` (#87)
* Move '//index/internal/collections' to '//internal/collections'
* Format BUILD files and update MODULE.bazel.lock
* Fix gazelle/go config for index/conan/integraiton_tests
