+++ BUILD.bazel
@@ -1,14 +1,17 @@
 load("@gazelle//:def.bzl", "gazelle", "gazelle_binary")
 load("@package_metadata//licenses/rules:license.bzl", "license")
-load("@package_metadata//rules:package_metadata.bzl", "package_metadata")
 load("@package_metadata//purl:purl.bzl", "purl")
+load("@package_metadata//rules:package_metadata.bzl", "package_metadata")
 
 package_metadata(
     name = "package_metadata",
     attributes = [
         ":license",
     ],
-    purl = purl.bazel(module_name(), module_version()),
+    purl = purl.bazel(
+        module_name(),
+        module_version(),
+    ),
     visibility = ["//visibility:public"],
 )
