+++ index/BUILD.bazel
@@ -2,8 +2,8 @@ test_suite(
     name = "integration_tests",
     tags = ["manual"],
     tests = [
-        "//index/internal/tests/example_integration_test",
         "//index/conan/integration_tests:integration_test",
+        "//index/internal/tests/example_integration_test",
         "//index/rules_foreign_cc/integration_tests:integration_test",
     ],
 )
