+++ index/conan/BUILD.bazel
@@ -9,9 +9,9 @@ go_library(
         "//index/conan/internal/targets",
         "//index/internal/bazel",
         "//index/internal/bazel/proto:build_go_proto",
-        "//index/internal/collections",
         "//index/internal/indexer",
         "//index/internal/indexer/cli",
+        "//internal/collections",
         "@gazelle//label",
     ],
 )
