+++ index/internal/indexer/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/EngFlow/gazelle_cc/index/internal/indexer",
     visibility = ["//index:__subpackages__"],
     deps = [
-        "//index/internal/collections",
+        "//internal/collections",
         "@gazelle//label",
     ],
 )
@@ -16,7 +16,7 @@ go_test(
     srcs = ["indexer_test.go"],
     embed = [":indexer"],
     deps = [
-        "//index/internal/collections",
+        "//internal/collections",
         "@com_github_stretchr_testify//assert",
         "@gazelle//label",
     ],
