+++ index/internal/indexer/indexer_test.go
@@ -18,7 +18,7 @@ import (
 	"log"
 	"testing"
 
-	"github.com/EngFlow/gazelle_cc/index/internal/collections"
+	"github.com/EngFlow/gazelle_cc/internal/collections"
 	"github.com/bazelbuild/bazel-gazelle/label"
 	"github.com/stretchr/testify/assert"
 )
