+++ index/rules_foreign_cc/BUILD.bazel
@@ -8,9 +8,9 @@ go_library(
     deps = [
         "//index/internal/bazel",
         "//index/internal/bazel/proto:build_go_proto",
-        "//index/internal/collections",
         "//index/internal/indexer",
         "//index/internal/indexer/cli",
+        "//internal/collections",
         "@gazelle//label",
     ],
 )
