+++ internal/collections/set.go
