+++ internal/collections/set_test.go
